backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
pydantic
python-multipart
requests
python-dotenv
fastapi[all]
